# swagger.yaml
openapi: 3.0.3
info:
  title: API de Monitoreo de Calidad del Aire
  description: API REST para el sistema de monitoreo de calidad del aire con arquitectura hexagonal
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: soporte@monitoreo-aire.com
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
security:
  - bearerAuth: []
  - deviceToken: []
paths:
  /register:
    post:
      summary: Registrar nuevo usuario
      description: Crea una nueva cuenta de usuario en el sistema
      tags:
        - Autenticación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
                - correo
                - clave
              properties:
                nombre:
                  type: string
                  example: "Juan Pérez"
                correo:
                  type: string
                  format: email
                  example: "juan.perez@email.com"
                clave:
                  type: string
                  minLength: 8
                  example: "password123"
                rol:
                  type: string
                  enum: [administrador, conserje]
                  default: conserje
                  example: "conserje"
      responses:
        201:
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        409:
          description: El correo ya está registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        400:
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /devices:
    post:
      summary: Crear dispositivo
      description: Registra un nuevo dispositivo de monitoreo (solo administradores)
      tags:
        - Dispositivos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
                - ubicacion
                - id_edificio
              properties:
                nombre:
                  type: string
                  example: "Sensor Piso 1 - Sala A"
                ubicacion:
                  type: string
                  example: "Edificio Principal, Piso 1, Sala de Reuniones A"
                id_edificio:
                  type: integer
                  example: 1
      responses:
        201:
          description: Dispositivo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispositivo'
        403:
          description: Permisos insuficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Obtener dispositivos
      description: Lista todos los dispositivos del sistema
      tags:
        - Dispositivos
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de dispositivos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dispositivo'
  /devices/{deviceId}:
    get:
      summary: Obtener dispositivo específico
      description: Obtiene la información de un dispositivo por su ID
      tags:
        - Dispositivos
      security:
        - bearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: Información del dispositivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispositivo'
        404:
          description: Dispositivo no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Pérez"
        correo:
          type: string
          format: email
          example: "juan.perez@email.com"
        rol:
          type: string
          enum: [administrador, conserje]
          example: "conserje"
        numero_contacto:
          type: string
          example: "+56912345678"
        url_imagen_perfil:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        id_edificio_asignado:
          type: integer
          example: 1
        fecha_creacion:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
    Dispositivo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Sensor Piso 1 - Sala A"
        ubicacion:
          type: string
          example: "Edificio Principal, Piso 1, Sala de Reuniones A"
        token_acceso:
          type: string
          example: "abc123def456ghi789"
        id_edificio:
          type: integer
          example: 1
        estado:
          type: string
          enum: [activo, inactivo, mantenimiento]
          example: "activo"
        fecha_instalacion:
          type: string
          format: date
          example: "2025-01-15"
        ultima_lectura:
          type: string
          format: date-time
          example: "2025-01-15T14:30:00Z"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Mensaje de error descriptivo"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /login

    deviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Token único del dispositivo para autenticación de IoT